% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ensemble.R
\name{simple_model_forecasts}
\alias{simple_model_forecasts}
\title{Makes a list of a set of simple models forecasts}
\usage{
simple_model_forecasts(ts, hor)
}
\arguments{
\item{ts}{Time series to forecast on}

\item{hor}{Forecast horizon}
}
\description{
This function creates a set of forecasts getting from the simple models
}
\examples{
simple_model_forecasts(ts, 1)
}
\keyword{forecast}
