% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_ensemble.R
\name{weighted_ensemble_forecast}
\alias{weighted_ensemble_forecast}
\title{Clculate forecast using weighted ensemble method}
\usage{
weighted_ensemble_forecast(qty_ts, N, forecast_horizon)
}
\arguments{
\item{qty_ts}{Time series to forecast on}

\item{N}{Number of holdouts to test simple models to define weights}

\item{forecast_horizon}{Forecast horizon}
}
\description{
This function calculates forecast basing on the weighted ensemble of simple methods from the function simple_model_forecasts
}
\examples{
weighted_ensemble_forecast(ts, 3, 18)
}
\keyword{forecast}
